refactor: share common build steps
